+++ minix/servers/ipc/shm.c
@@ -248,8 +248,8 @@ static void fill_shminfo(struct shminfo * sinfo)
 	sinfo->shmmax = (unsigned long)-1;
 	sinfo->shmmin = 1;
 	sinfo->shmmni = __arraycount(shm_list);
-	sinfo->shmseg = (unsigned long)-1;
-	sinfo->shmall = (unsigned long)-1;
+	sinfo->shmseg = (uint32_t)-1;
+	sinfo->shmall = (uint32_t)-1;
 }
 
 int do_shmctl(message * m)
